move fixed size list to a seperate file
