+++ arrow/src/array/mod.rs
@@ -164,6 +164,7 @@ mod array_boolean;
 mod array_decimal;
 mod array_dictionary;
 mod array_fixed_size_binary;
+mod array_fixed_size_list;
 mod array_list;
 mod array_map;
 mod array_primitive;
@@ -202,12 +203,12 @@ pub use self::array_decimal::BasicDecimalArray;
 pub use self::array_decimal::Decimal128Array;
 pub use self::array_decimal::Decimal256Array;
 pub use self::array_fixed_size_binary::FixedSizeBinaryArray;
+pub use self::array_fixed_size_list::FixedSizeListArray;
 
 #[deprecated(note = "Please use `Decimal128Array` instead")]
 pub type DecimalArray = Decimal128Array;
 
 pub use self::array_dictionary::DictionaryArray;
-pub use self::array_list::FixedSizeListArray;
 pub use self::array_list::LargeListArray;
 pub use self::array_list::ListArray;
 pub use self::array_map::MapArray;
